requirements.txt
Flask
openai
python-docx
requests
python-dotenv
